Fixes for using size_t/ssize_t/su_socket_t and isize_t/usize_t/issize_t.
su, sip, msg, http, url, sdp, soa, etc..

darcs-hash:20060919133329-65a35-62b9f9819398d39e889315e8bd1b24ba0b8a9b23.gz

+++ libsofia-sip-ua/tport/tport_internal.h
@@ -394,10 +394,10 @@ void tport_recv_event(tport_t *self);
 void tport_send_event(tport_t *self);
 void tport_hup_event(tport_t *self);
 
-int tport_recv_iovec(tport_t const *self, 
-		     msg_t **mmsg,
-		     msg_iovec_t iovec[msg_n_fragments], int N, 
-		     int exact);
+ssize_t tport_recv_iovec(tport_t const *self, 
+			 msg_t **mmsg,
+			 msg_iovec_t iovec[msg_n_fragments], int N, 
+			 int exact);
 
 msg_t *tport_msg_alloc(tport_t const *self, unsigned size);
